+++ components/amorphic/lib/utils/InputValidator.ts
@@ -3,6 +3,8 @@ import { Request, Response } from 'express';
 import * as AmorphicContext from '../AmorphicContext';
 import { SupertypeSession } from '@haventech/supertype';
 
+const moduleName = `amorphic/lib/utils/InputValidator`;
+
 export class InputValidator {
     /**
      * Validate URL parameters
@@ -62,6 +64,7 @@ export class InputValidator {
     }
 
     private static logAndCounterValue(originalValue: string, validatedValue: string, validatorLog: boolean, counterField: string, additionalLogs) {
+        const functionName = this.logAndCounterValue.name;
         if (validatedValue !== originalValue) {
             const statsdClient = SupertypeSession.statsdClient;
             if(statsdClient
@@ -70,16 +73,17 @@ export class InputValidator {
                 statsdClient.counter(`amorphic.server.validator.${counterField}.counter`, 1);
             }
             if (validatorLog) {
-                SupertypeSession.logger.info(
-                    {
-                        component: 'amorphic',
-                        module: 'InputValidator',
+                SupertypeSession.logger.info({
+                    module: moduleName,
+                    function: functionName,
+                    category: 'milestone',
+                    data: {
                         activity: 'performValidations',
                         originalValue: originalValue,
                         validatedValue: validatedValue,
-                        ...additionalLogs
+                        additionalLogs: additionalLogs
                     }
-                );
+                });
             }
             return validatedValue;
         }