+++ components/supertype/HISTORY.md
@@ -1,3 +1,7 @@
+## 6.0.0
+* Upgrade Supertype logger to use ClientLogger (client passed logger).
+## 5.1.0
+* Allow setting of logger on ObjectTemplate
 ## 5.0.1
 * bump nconf version due to security vulnerability
 ## 5.0.0